refactor: modularizacao + mods de funcoes aux
