+++ about.h
@@ -1,0 +1,2 @@
+void about_soft();
+void about_dev();
