+++ animal.h
@@ -1,0 +1,6 @@
+// Menu animal
+char animal_menu();
+void create_animal();
+void search_animal();
+void edit_animal();
+void delete_animal();
