+++ aux_functions.h
@@ -1,0 +1,3 @@
+void header();
+void header_complete();
+void no_operation();
