+++ client.h
@@ -1,0 +1,6 @@
+// Menu cliente
+char client_menu();
+void create_client();
+void search_client();
+void edit_client();
+void delete_client();
