+++ product.h
@@ -1,0 +1,6 @@
+// Menu produto
+char product_menu();
+void create_product();
+void search_product();
+void edit_product();
+void delete_product();
