+++ sale.h
@@ -1,0 +1,5 @@
+// Menu venda  
+char sale_menu();
+void create_sale();
+void search_sale();
+void delete_sale();
