+++ worker.h
@@ -1,0 +1,6 @@
+// Menu funcionário
+char worker_menu();
+void create_worker();
+void search_worker();
+void edit_worker();
+void delete_worker();
